+++ backend/src/main/java/com/woowacourse/gongseek/auth/application/dto/AccessTokenResponse.java
@@ -1,4 +1,4 @@
-package com.woowacourse.gongseek.auth.presentation.dto;
+package com.woowacourse.gongseek.auth.application.dto;
 
 import lombok.AccessLevel;
 import lombok.AllArgsConstructor;
