+++ backend/src/main/java/com/woowacourse/gongseek/auth/application/dto/AppMember.java
@@ -1,4 +1,4 @@
-package com.woowacourse.gongseek.auth.presentation.dto;
+package com.woowacourse.gongseek.auth.application.dto;
 
 import lombok.AccessLevel;
 import lombok.AllArgsConstructor;
@@ -8,7 +8,9 @@ import lombok.NoArgsConstructor;
 @AllArgsConstructor
 @NoArgsConstructor(access = AccessLevel.PRIVATE)
 @Getter
-public class OAuthLoginUrlResponse {
+public abstract class AppMember {
 
-    private String url;
+    private Long payload;
+
+    public abstract boolean isGuest();
 }
