+++ backend/src/main/java/com/woowacourse/gongseek/auth/application/dto/OAuthCodeRequest.java
@@ -1,5 +1,6 @@
-package com.woowacourse.gongseek.auth.presentation.dto;
+package com.woowacourse.gongseek.auth.application.dto;
 
+import javax.validation.constraints.NotNull;
 import lombok.AccessLevel;
 import lombok.AllArgsConstructor;
 import lombok.Getter;
@@ -8,9 +9,8 @@ import lombok.NoArgsConstructor;
 @AllArgsConstructor
 @NoArgsConstructor(access = AccessLevel.PRIVATE)
 @Getter
-public abstract class AppMember {
+public class OAuthCodeRequest {
 
-    private Long payload;
-
-    public abstract boolean isGuest();
+    @NotNull
+    private String code;
 }
