+++ backend/src/main/java/com/woowacourse/gongseek/auth/application/dto/OAuthLoginUrlResponse.java
@@ -1,6 +1,5 @@
-package com.woowacourse.gongseek.auth.presentation.dto;
+package com.woowacourse.gongseek.auth.application.dto;
 
-import javax.validation.constraints.NotNull;
 import lombok.AccessLevel;
 import lombok.AllArgsConstructor;
 import lombok.Getter;
@@ -9,8 +8,7 @@ import lombok.NoArgsConstructor;
 @AllArgsConstructor
 @NoArgsConstructor(access = AccessLevel.PRIVATE)
 @Getter
-public class OAuthCodeRequest {
+public class OAuthLoginUrlResponse {
 
-    @NotNull
-    private String code;
+    private String url;
 }
