+++ backend/src/test/java/com/woowacourse/gongseek/acceptance/CommentAcceptanceTest.java
@@ -13,7 +13,7 @@ import static org.assertj.core.api.Assertions.assertThat;
 import static org.junit.jupiter.api.Assertions.assertAll;
 
 import com.woowacourse.gongseek.article.application.dto.ArticleIdResponse;
-import com.woowacourse.gongseek.auth.presentation.dto.AccessTokenResponse;
+import com.woowacourse.gongseek.auth.application.dto.AccessTokenResponse;
 import com.woowacourse.gongseek.comment.presentation.dto.CommentResponse;
 import com.woowacourse.gongseek.common.exception.dto.ErrorResponse;
 import io.restassured.response.ExtractableResponse;
