+++ backend/src/test/java/com/woowacourse/gongseek/acceptance/support/FakeGithubController.java
@@ -1,8 +1,8 @@
 package com.woowacourse.gongseek.acceptance.support;
 
-import com.woowacourse.gongseek.auth.presentation.dto.GithubAccessTokenRequest;
-import com.woowacourse.gongseek.auth.presentation.dto.GithubAccessTokenResponse;
-import com.woowacourse.gongseek.auth.presentation.dto.GithubProfileResponse;
+import com.woowacourse.gongseek.auth.application.dto.GithubAccessTokenRequest;
+import com.woowacourse.gongseek.auth.application.dto.GithubAccessTokenResponse;
+import com.woowacourse.gongseek.auth.application.dto.GithubProfileResponse;
 import com.woowacourse.gongseek.auth.support.GithubClientFixtures;
 import org.springframework.http.HttpHeaders;
 import org.springframework.http.ResponseEntity;
