+++ backend/src/test/java/com/woowacourse/gongseek/acceptance/support/fixtures/AuthFixture.java
@@ -1,8 +1,8 @@
 package com.woowacourse.gongseek.acceptance.support.fixtures;
 
-import com.woowacourse.gongseek.auth.presentation.dto.AccessTokenResponse;
-import com.woowacourse.gongseek.auth.presentation.dto.OAuthCodeRequest;
-import com.woowacourse.gongseek.auth.presentation.dto.OAuthLoginUrlResponse;
+import com.woowacourse.gongseek.auth.application.dto.AccessTokenResponse;
+import com.woowacourse.gongseek.auth.application.dto.OAuthCodeRequest;
+import com.woowacourse.gongseek.auth.application.dto.OAuthLoginUrlResponse;
 import com.woowacourse.gongseek.auth.support.GithubClientFixtures;
 import io.restassured.RestAssured;
 import io.restassured.response.ExtractableResponse;
