+++ backend/src/test/java/com/woowacourse/gongseek/acceptance/support/fixtures/LikeFixture.java
@@ -1,7 +1,7 @@
 package com.woowacourse.gongseek.acceptance.support.fixtures;
 
 import com.woowacourse.gongseek.article.application.dto.ArticleIdResponse;
-import com.woowacourse.gongseek.auth.presentation.dto.AccessTokenResponse;
+import com.woowacourse.gongseek.auth.application.dto.AccessTokenResponse;
 import io.restassured.RestAssured;
 import io.restassured.response.ExtractableResponse;
 import io.restassured.response.Response;
