+++ backend/src/test/java/com/woowacourse/gongseek/acceptance/support/fixtures/VoteFixture.java
@@ -1,6 +1,6 @@
 package com.woowacourse.gongseek.acceptance.support.fixtures;
 
-import com.woowacourse.gongseek.auth.presentation.dto.AccessTokenResponse;
+import com.woowacourse.gongseek.auth.application.dto.AccessTokenResponse;
 import com.woowacourse.gongseek.vote.presentation.dto.SelectVoteItemIdRequest;
 import com.woowacourse.gongseek.vote.presentation.dto.VoteCreateRequest;
 import io.restassured.RestAssured;
