+++ backend/src/test/java/com/woowacourse/gongseek/auth/application/AuthServiceTest.java
@@ -7,11 +7,11 @@ import static org.assertj.core.api.Assertions.assertThatThrownBy;
 import static org.junit.jupiter.api.Assertions.assertAll;
 import static org.mockito.BDDMockito.given;
 
+import com.woowacourse.gongseek.auth.application.dto.GithubProfileResponse;
+import com.woowacourse.gongseek.auth.application.dto.TokenResponse;
 import com.woowacourse.gongseek.auth.domain.RefreshToken;
 import com.woowacourse.gongseek.auth.domain.repository.RefreshTokenRepository;
 import com.woowacourse.gongseek.auth.exception.InvalidRefreshTokenException;
-import com.woowacourse.gongseek.auth.presentation.dto.GithubProfileResponse;
-import com.woowacourse.gongseek.auth.presentation.dto.TokenResponse;
 import com.woowacourse.gongseek.member.domain.Member;
 import com.woowacourse.gongseek.member.domain.repository.MemberRepository;
 import com.woowacourse.gongseek.support.AuthIntegrationTest;
