+++ backend/src/test/java/com/woowacourse/gongseek/auth/presentation/AuthControllerTest.java
@@ -16,8 +16,8 @@ import static org.springframework.restdocs.payload.PayloadDocumentation.response
 import static org.springframework.test.web.servlet.result.MockMvcResultHandlers.print;
 import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.status;
 
-import com.woowacourse.gongseek.auth.presentation.dto.OAuthCodeRequest;
-import com.woowacourse.gongseek.auth.presentation.dto.TokenResponse;
+import com.woowacourse.gongseek.auth.application.dto.OAuthCodeRequest;
+import com.woowacourse.gongseek.auth.application.dto.TokenResponse;
 import com.woowacourse.gongseek.support.ControllerTest;
 import java.util.UUID;
 import org.junit.jupiter.api.DisplayName;
